Fixes different simulation outputs for different gc intervals.
Previously, the parent nodes of the next generation were being reverse sorted, so of course when I changed how often I simplified, different would-be parents would get reverse sorted in the tables.

Not a bug in the sense that the simulation output is still a valid simulation. However, different gc intervals producing different results is bad for consistency. New version preserves the parent order from the forward simulation and appends the ancestral samples in the back.